separate tasks api from regular endpoints
