+++ src/bptl/dashboard/urls.py
@@ -1,6 +1,6 @@
-from django.urls import path
+from django.urls import include, path
 
-from .views import AggregateView, TaskDetailView, TaskListView
+from .views import TaskDetailView, TaskListView
 
 app_name = "dashboard"
 
@@ -8,5 +8,5 @@ urlpatterns = [
     # Simply show the master template.
     path("", TaskListView.as_view(), name="task-list"),
     path("<pk>/", TaskDetailView.as_view(), name="task-detail"),
-    path("aggregate/", AggregateView.as_view(), name="aggregate"),
+    path("api/", include("bptl.dashboard.api.urls")),
 ]
